Removed old PetTimeline component
