--- src/parser/warlock/demonology/modules/pets/PetTimelineTab/TabComponent/SpellTimeline.css
@@ -1,164 +1,0 @@
-.spell-timeline .timeline {
-  background: rgba(12, 22, 29, 0.79);
-  box-shadow: inset 0 0 16px 3px #000;
-  border: 1px solid #000;
-  width: 100%;
-  height: 68px;
-  overflow: auto;
-}
-.spell-timeline .lane {
-  height: 28px;
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-}
-.spell-timeline .legend {
-  min-width: 200px;
-}
-@media (max-width: 800px) {
-  .spell-timeline .legend {
-    min-width: auto;
-    max-width: 100px;
-  }
-}
-.spell-timeline .legend .lane {
-  width: 100%;
-  padding-top: 1px;
-  padding-left: 6px;
-  padding-right: 15px;
-  min-width: auto;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-.spell-timeline .legend img {
-  height: 23px;
-}
-.spell-timeline .ruler {
-  white-space: nowrap;
-  height: 33px;
-  margin-bottom: 3px;
-}
-.spell-timeline .ruler-lane {
-  border-bottom: 1px solid #fff;
-  height: 36px;
-}
-.spell-timeline .ruler > * {
-  display: inline-block;
-  border-left: 1px solid #fff;
-  border-bottom: 1px solid #fff;
-  padding-left: 3px;
-  line-height: 1;
-  font-size: 12px;
-  position: relative;
-}
-.spell-timeline .ruler > *:before, .spell-timeline .ruler > *:after {
-  content: '';
-  position: absolute;
-  left: 50%;
-  bottom: 0;
-  height: 8px;
-  border-left: 1px solid #fff;
-}
-.spell-timeline .ruler.interval-3 > *:before {
-  left: 33%;
-}
-.spell-timeline .ruler.interval-3 > *:after {
-  left: 66%;
-}
-.spell-timeline .ruler:not(.interval-3) > *:after {
-  display: none;
-}
-
-.spell-timeline .events {
-  position: relative;
-  height: 2em;
-}
-.spell-timeline .events > * {
-  position: absolute;
-  display: block;
-  top: 1px;
-  height: 25px;
-  z-index: 1;
-}
-.spell-timeline .events > * a {
-  display: inline-block;
-}
-
-.spell-timeline .cast {
-  z-index: 10;
-  top: -1px;
-}
-.spell-timeline .cast img {
-  height: 23px;
-  margin-top: 0;
-}
-.spell-timeline .cast.with-buff {
-  top: 0px;
-}
-.spell-timeline .cast.with-buff img {
-  height: 21px;
-}
-.spell-timeline .cooldown {
-  top: 2px;
-  height: 23px;
-  box-shadow: 0 0 0 rgba(150, 150, 150, 0.4), 0 0 6px rgba(150, 150, 150, 0.4);
-  background: rgba(150, 150, 150, 0.4);
-}
-
-.spell-timeline .events .casting-time {
-  background: rgba(120, 102, 255, 0.4);
-  border-right: 2px solid rgba(250, 250, 250, 0.6);
-}
-.spell-timeline .events .inefficient img {
-  border: 2px solid red;
-  box-sizing: content-box;
-  margin-left: -1px;
-  margin-top: -1px;
-  box-shadow: 0 0 11px red;
-  border-radius: 2px;
-}
-
-.spell-timeline .events .enhanced img {
-  border: 2px solid rgb(226, 226, 226);
-  box-sizing: content-box;
-  margin-left: -1px;
-  margin-top: -1px;
-  box-shadow: 0 0 11px rgb(173, 173, 173);
-  border-radius: 2px;
-}
-.spell-timeline .death, .spell-timeline .resurrection {
-  position: absolute;
-  top: 0;
-  height: 100%;
-  width: 2px;
-  background: red;
-  box-shadow: 0 0 11px red;
-}
-.spell-timeline .resurrection {
-  background: lime;
-  box-shadow: 0 0 11px lime;
-}
-
-/* vertical scrollbar track */
-.gm-scrollbar.-vertical {
-  display: none;
-}
-
-/* horizontal scrollbar track */
-.gm-scrollbar.-horizontal {
-  background-color: rgba(50, 52, 53, 1);
-  border-radius: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  height: 9px;
-}
-
-/* scrollbar thumb */
-.gm-scrollbar .thumb {
-  background-color: #F8F8F8;
-  box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.2), 1px 0 1px rgba(0, 0, 0, 0.5);
-  border-radius: 5px;
-}
-.gm-scrollbar .thumb:hover, .gm-scrollbar .thumb:active {
-  background-color: #dedede;
-}
