+++ src/parser/warlock/demonology/modules/pets/PetTimelineTab/TabComponent/Timeline.scss
